// .angular/src/app/models/TOC/productosAdelanto.interface.ts
export interface ProductoAdelanto {
    Nombre:              string;
    Direccion:           string;
    Ciudad:              string;
    Telefono:            string;
    entrega:             string;
    Descripcion:         string;
    Fecha_comp_original: string;
    Fecha_reprogramada:  string | null;
    Region:              string;
    Observacion:         string
}